Extracts presence related logic into presence.ts
