+++ assets/src/components/conversations/ConversationsProvider.tsx
@@ -11,6 +11,11 @@ import {
 } from '../../utils';
 import {SOCKET_URL} from '../../socket';
 import logger from '../../logger';
+import {
+  PhoenixPresence,
+  PresenceDiff,
+  updatePresenceWithDiff,
+} from '../../presence';
 
 export const ConversationsContext = React.createContext<{
   loading: boolean;
@@ -72,82 +77,6 @@ export const ConversationsContext = React.createContext<{
 export const useConversations = () => useContext(ConversationsContext);
 
 type ConversationBucket = 'all' | 'mine' | 'priority' | 'closed';
-
-type PresenceMetadata = {online_at?: string; phx_ref: string};
-type PhoenixPresence = {
-  [key: string]: {
-    metas: Array<PresenceMetadata>;
-  } | null;
-};
-type PresenceDiff = {
-  joins: PhoenixPresence;
-  leaves: PhoenixPresence;
-};
-
-export const updatePresenceWithJoiners = (
-  joiners: PhoenixPresence,
-  currentState: PhoenixPresence
-): PhoenixPresence => {
-  // Update our presence state by adding all the joiners, represented by
-  // keys like "customer:1a2b3c", "user:123", etc.
-  // The `metas` represent the metadata of each presence. A single user/customer
-  // can have multiple `metas` if logged into multiple devices/windows.
-  let result = {...currentState};
-
-  Object.keys(joiners).forEach((key) => {
-    const existing = result[key];
-    const update = joiners[key];
-
-    // `metas` is how Phoenix tracks each individual presence
-    if (!update || !update.metas) {
-      throw new Error(`Unexpected join state: ${update}`);
-    }
-
-    if (existing && existing.metas) {
-      result[key] = {metas: [...existing.metas, ...update.metas]};
-    } else {
-      result[key] = {metas: update.metas};
-    }
-  });
-
-  return result;
-};
-
-export const updatePresenceWithExiters = (
-  exiters: PhoenixPresence,
-  currentState: PhoenixPresence
-): PhoenixPresence => {
-  // Update our presence state by removing all the exiters, represented by
-  // keys like "customer:1a2b3c", "user:123", etc. We currently indicate an
-  // "exit" by setting their key to `null`.
-  // The `metas` represent the metadata of each presence. A single user/customer
-  // can have multiple `metas` if logged into multiple devices/windows.
-  let result = {...currentState};
-
-  Object.keys(exiters).forEach((key) => {
-    const existing = result[key];
-    const update = exiters[key];
-
-    // `metas` is how Phoenix tracks each individual presence
-    if (!update || !update.metas) {
-      throw new Error(`Unexpected leave state: ${update}`);
-    }
-
-    if (existing && existing.metas) {
-      const remaining = existing.metas.filter((meta: PresenceMetadata) => {
-        return update.metas.some(
-          (m: PresenceMetadata) => meta.phx_ref !== m.phx_ref
-        );
-      });
-
-      result[key] = remaining.length ? {metas: remaining} : null;
-    } else {
-      result[key] = null;
-    }
-  });
-
-  return result;
-};
 
 type Props = React.PropsWithChildren<{}>;
 type State = {
@@ -286,21 +215,9 @@ export class ConversationsProvider extends React.Component<Props, State> {
   };
 
   handlePresenceDiff = (diff: PresenceDiff) => {
-    const {joins, leaves} = diff;
-    const {presence} = this.state;
-
-    const withJoins = updatePresenceWithJoiners(joins, presence);
-    const withLeaves = updatePresenceWithExiters(leaves, presence);
-    const combined = {...withJoins, ...withLeaves};
-    const latest = Object.keys(combined).reduce((acc, key: string) => {
-      if (!combined[key]) {
-        return acc;
-      }
-
-      return {...acc, [key]: combined[key]};
-    }, {} as PhoenixPresence);
-
-    this.setState({presence: latest});
+    this.setState({
+      presence: updatePresenceWithDiff(this.state.presence, diff),
+    });
   };
 
   isCustomerOnline = (customerId: string) => {
